+++ src/nectarbase/__init__.py
@@ -3,6 +3,7 @@
 from .version import version as __version__
 
 __all__ = [
+    "__version__",
     "memo",
     "objects",
     "objecttypes",