+++ src/nectargraphenebase/operations.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+from collections import OrderedDict
+
 from .objects import GrapheneObject, isArgsThisClass
 from .types import (
     Set,